Changed All Callback functions to include aliasing

+++ SOAP/SOAP/Models/Callbacks/AnestheticPlanInjectionCallback.cs
@@ -12,22 +12,22 @@ namespace SOAP.Models.Callbacks
             AnestheticPlanInjection anesPlanInject = new AnestheticPlanInjection();
             anesPlanInject.Id = Convert.ToInt32(read["a.Id"]);
             anesPlanInject.PatientId = Convert.ToInt32(read["a.PatientId"].ToString());
-            anesPlanInject.Drug.Drug.Id = Convert.ToInt16(read["a.DrugId"].ToString());
-            anesPlanInject.Route.Id = Convert.ToInt16(read["a.Route_Id"].ToString());
+            anesPlanInject.Drug.Drug.Id = Convert.ToInt32(read["a.DrugId"].ToString());
+            anesPlanInject.Route.Id = Convert.ToInt32(read["a.Route_Id"].ToString());
             anesPlanInject.Dosage = Convert.ToDecimal(read["a.Dosage"].ToString());
 
             foreach (AnestheticPlanInjection.LazyComponents a in lazyComponents)
             {
                 if (a == AnestheticPlanInjection.LazyComponents.LOAD_ROUTE_WITH_DETAILS)
                 {
-                    anesPlanInject.Route.Category.Id = Convert.ToInt16(read["c.CategoryId"].ToString());
+                    anesPlanInject.Route.Category.Id = Convert.ToInt32(read["c.CategoryId"].ToString());
                     anesPlanInject.Route.Label = read["c.Label"].ToString();
                     anesPlanInject.Route.OtherFlag = Convert.ToChar(read["c.OtherFlag"].ToString());
                     anesPlanInject.Route.Label = read["c.Description"].ToString();
                 }
                 else if (a == AnestheticPlanInjection.LazyComponents.LOAD_DRUG_INFORMATION)
                 {
-                    anesPlanInject.Drug.Drug.Category.Id = Convert.ToInt16(read["d.CategoryId"].ToString());
+                    anesPlanInject.Drug.Drug.Category.Id = Convert.ToInt32(read["d.CategoryId"].ToString());
                     anesPlanInject.Drug.Drug.Label = read["d.Label"].ToString();
                     anesPlanInject.Drug.Drug.OtherFlag = Convert.ToChar(read["d.OtherFlag"].ToString());
                     anesPlanInject.Drug.Drug.Description = read["b.Description"].ToString();
